Reorganise test mocks
