+++ apps/ui/test/application.mocks.ts
@@ -5,35 +5,17 @@ import {
   type ApplicationRoundNode,
   ApplicationRoundReservationCreationStatusChoice,
   ApplicationRoundStatusChoice,
-  type ApplicationRoundTimeSlotNode,
   ApplicationSectionStatusChoice,
   ApplicationStatusChoice,
   type ApplicationViewFragment,
-  Authentication,
   CreateApplicationDocument,
   type CreateApplicationMutation,
   type CreateApplicationMutationVariables,
-  CurrentUserDocument,
-  type CurrentUserQuery,
-  ImageType,
-  OptionsDocument,
-  type OptionsQuery,
   OrganizationTypeChoice,
   Priority,
   type PurposeNode,
-  ReservationKind,
-  ReservationPurposeOrderingChoices,
-  ReservationStartInterval,
   type ReservationUnitNode,
-  ReservationUnitOrderingChoices,
-  ReservationUnitPublishingState,
-  ReservationUnitReservationState,
-  ReservationUnitTypeOrderingChoices,
-  SearchReservationUnitsDocument,
-  type SearchReservationUnitsQuery,
-  type SearchReservationUnitsQueryVariables,
   TermsType,
-  type UnitNode,
   UpdateApplicationDocument,
   type UpdateApplicationMutation,
   Weekday,
@@ -41,29 +23,13 @@ import {
 import { base64encode } from "common/src/helpers";
 import { addDays, addMonths, addYears } from "date-fns";
 import {
-  createMockReservationUnitType,
-  createOptionQueryMock,
-  generateDescriptionFragment,
   generateNameFragment,
   generateTextFragment,
-  type ICreateGraphQLMock,
-  type CreateGraphQLMockProps,
   type CreateGraphQLMocksReturn,
 } from "./test.gql.utils";
+import { createMockReservationUnit } from "./reservation-unit.mocks";
 
-export function createGraphQLMocks({
-  noUser = false,
-  isSearchError = false,
-}: CreateGraphQLMockProps = {}): CreateGraphQLMocksReturn {
-  return [
-    ...createSearchQueryMocks({ isSearchError }),
-    ...createOptionsQueryMocks(),
-    ...createCurrentUserQueryMocks({ noUser }),
-    ...createApplicationMutationMocks(),
-  ];
-}
-
-function createApplicationMutationMocks(): CreateGraphQLMocksReturn {
+export function createApplicationMutationMocks(): CreateGraphQLMocksReturn {
   const createVariables: CreateApplicationMutationVariables = {
     input: {
       applicationRound: 1,
@@ -100,312 +66,6 @@ function createApplicationMutationMocks(): CreateGraphQLMocksReturn {
       },
     },
   ];
-}
-
-interface CurrentUserQueryMocksProps extends ICreateGraphQLMock {
-  noUser: boolean;
-}
-
-function createCurrentUserQueryMocks({
-  noUser,
-}: CurrentUserQueryMocksProps): CreateGraphQLMocksReturn {
-  const CurrentUserMock: CurrentUserQuery = {
-    currentUser: !noUser
-      ? {
-          id: base64encode("UserNode:1"),
-          pk: 1,
-          firstName: "Test",
-          lastName: "User",
-          email: "test@user",
-          isAdAuthenticated: false,
-        }
-      : null,
-  };
-
-  return [
-    {
-      request: {
-        query: CurrentUserDocument,
-      },
-      result: {
-        data: CurrentUserMock,
-      },
-    },
-  ];
-}
-
-function createOptionsQueryMocks(): CreateGraphQLMocksReturn {
-  const OptionsMock: OptionsQuery = createOptionQueryMock();
-  return [
-    {
-      request: {
-        query: OptionsDocument,
-        variables: {
-          reservationUnitTypesOrderBy:
-            ReservationUnitTypeOrderingChoices.RankAsc,
-          reservationPurposesOrderBy: ReservationPurposeOrderingChoices.RankAsc,
-          unitsOrderBy: [],
-          equipmentsOrderBy: [],
-          purposesOrderBy: [],
-        },
-      },
-      result: {
-        data: OptionsMock,
-      },
-    },
-  ];
-}
-
-interface SearchQueryProps extends ICreateGraphQLMock {
-  isSearchError: boolean;
-}
-function createSearchQueryMocks({
-  isSearchError,
-}: SearchQueryProps): CreateGraphQLMocksReturn {
-  // TODO this should enforce non nullable for the query
-  // it can be null when the query is loading, but when we mock it it should be non nullable
-  // Q: what about failed queries? (though they should have different type)
-  const SearchReservationUnitsQueryMock: SearchReservationUnitsQuery = {
-    reservationUnits: {
-      totalCount: 10,
-      edges: Array.from({ length: 10 }, (_, i) => i + 1).map((pk) => ({
-        node: createMockReservationUnit({ pk }),
-      })),
-      pageInfo: {
-        // TOOD how to mock this?
-        endCursor: null,
-        hasNextPage: false,
-      },
-    },
-  };
-  const SearchReservationUnitsQueryMockWithParams: SearchReservationUnitsQuery =
-    {
-      reservationUnits: {
-        totalCount: 1,
-        edges: [
-          {
-            node: createMockReservationUnit({ pk: 1 }),
-          },
-        ],
-        pageInfo: {
-          // TOOD how to mock this?
-          endCursor: null,
-          hasNextPage: false,
-        },
-      },
-    };
-
-  return [
-    {
-      request: {
-        query: SearchReservationUnitsDocument,
-        variables: createSearchVariablesMock(),
-      },
-      result: {
-        data: SearchReservationUnitsQueryMock,
-      },
-      // There are no different errors for this query result (just default error text)
-      error: isSearchError ? new Error("Search error") : undefined,
-    },
-    {
-      request: {
-        query: SearchReservationUnitsDocument,
-        variables: createSearchVariablesMock({ date: null }),
-      },
-      result: {
-        data: SearchReservationUnitsQueryMock,
-      },
-      error: isSearchError ? new Error("Search error") : undefined,
-    },
-    {
-      request: {
-        query: SearchReservationUnitsDocument,
-        variables: createSearchVariablesMock({ textSearch: "foobar" }),
-      },
-      result: {
-        data: SearchReservationUnitsQueryMockWithParams,
-      },
-      error: isSearchError ? new Error("Search error") : undefined,
-    },
-  ];
-}
-
-function createSearchVariablesMock({
-  textSearch = null,
-  date = new Date(2024, 1, 1),
-}: {
-  textSearch?: string | null;
-  date?: Date | null;
-  // TODO return type issues because all of them are optional (by backend)
-  // we'd need to match them to a Required return type that we actully use
-  // so what happens:
-  // a new query param is added but that is not reflected in the mock
-  // -> this is not a lint / type error but a runtime error in the tests
-} = {}): Readonly<SearchReservationUnitsQueryVariables> {
-  return {
-    textSearch,
-    purposes: [],
-    unit: [],
-    reservationUnitType: [],
-    equipments: [],
-    accessType: [],
-    accessTypeBeginDate: date ? date.toISOString() : null,
-    accessTypeEndDate: date ? addYears(date, 1).toISOString() : null,
-    reservableDateStart: date ? date.toISOString() : null,
-    reservableDateEnd: null,
-    reservableTimeStart: null,
-    reservableTimeEnd: null,
-    reservableMinimumDurationMinutes: null,
-    applicationRound: [1],
-    personsAllowed: null,
-    first: 36,
-    orderBy: [
-      ReservationUnitOrderingChoices.NameFiAsc,
-      ReservationUnitOrderingChoices.PkAsc,
-    ],
-    isDraft: false,
-    isVisible: true,
-    reservationKind: ReservationKind.Season,
-  } as const;
-}
-
-export function createMockReservationUnit({
-  pk,
-}: {
-  pk: number;
-}): ReservationUnitNode {
-  const timeSelector: ApplicationRoundTimeSlotNode = {
-    id: base64encode(`ApplicationRoundTimeSlotNode:1`),
-    pk,
-    weekday: 1,
-    closed: false,
-    reservableTimes: [
-      {
-        begin: "08:00",
-        end: "16:00",
-      },
-    ],
-  };
-  return {
-    id: base64encode(`ReservationUnitNode:${pk}`),
-    pk,
-    ...generateNameFragment(`ReservationUnit ${pk}`),
-    // TODO this is weird
-    reservationBegins: addYears(new Date(), -1 * pk).toISOString(),
-    reservationEnds: addYears(new Date(), 1 * pk).toISOString(),
-    isClosed: false,
-    // TODO implement though for Seasonal this doesn't matter
-    firstReservableDatetime: null,
-    currentAccessType: null,
-    effectiveAccessType: null,
-    maxPersons: null,
-    // TODO implement though for Seasonal this doesn't matter
-    pricings: [],
-    unit: createMockUnit({ pk }),
-    reservationUnitType: createMockReservationUnitType({
-      name: "ReservationUnitType",
-    }),
-    images: [
-      {
-        id: base64encode("Image:1"),
-        pk: 1,
-        imageUrl: "https://example.com/image1.jpg",
-        largeUrl: "https://example.com/image1_large.jpg",
-        mediumUrl: "https://example.com/image1_medium.jpg",
-        smallUrl: "https://example.com/image1_small.jpg",
-        imageType: ImageType.Main,
-      },
-    ] as const,
-    accessTypes: [],
-    // Everything below is only for completeness of the mock type (not used for application tests)
-    // TODO this can be removed
-    allowReservationsWithoutOpeningHours: false,
-    // applicationRoundTimeSlots: [] as const, // ReadonlyArray<ApplicationRoundTimeSlotNode>;
-    applicationRoundTimeSlots: [timeSelector],
-    applicationRounds: [] as const, // ReadonlyArray<ApplicationRoundNode>;
-    authentication: Authentication.Weak,
-    bufferTimeAfter: 0, //Scalars["Duration"]["output"];
-    bufferTimeBefore: 0, // Scalars["Duration"]["output"];
-    calculatedSurfaceArea: 0, // Scalars["Int"]["output"];
-    canApplyFreeOfCharge: false, // Scalars["Boolean"]["output"];
-    cancellationRule: null, // Maybe<ReservationUnitCancellationRuleNode>;
-    cancellationTerms: null, // Maybe<TermsOfUseNode>;
-    contactInformation: "", // admin-ui only feature
-    descriptionEn: "Description EN",
-    descriptionFi: "Description FI",
-    descriptionSv: "Description SV",
-    equipments: [] as const, // ReadonlyArray<EquipmentNode>;
-    haukiUrl: null, // Maybe<Scalars["String"]["output"]>;
-    isArchived: false,
-    isDraft: false,
-    location: null,
-    maxReservationDuration: null, // Maybe<Scalars["Duration"]["output"]>;
-    maxReservationsPerUser: null, //Maybe<Scalars["Int"]["output"]>;
-    metadataSet: null, //Maybe<ReservationMetadataSetNode>;
-    minPersons: null, //Maybe<Scalars["Int"]["output"]>;
-    minReservationDuration: null, // Maybe<Scalars["Duration"]["output"]>;
-    numActiveUserReservations: 0, // Scalars["Int"]["output"];
-    paymentMerchant: null, //Maybe<PaymentMerchantNode>;
-    paymentProduct: null, //Maybe<PaymentProductNode>;
-    paymentTerms: null, // Maybe<TermsOfUseNode>;
-    pricingTerms: null, //Maybe<TermsOfUseNode>;
-    publishBegins: null, // Maybe<Scalars["DateTime"]["output"]>;
-    publishEnds: null, // Maybe<Scalars["DateTime"]["output"]>;
-    publishingState: ReservationUnitPublishingState.Published,
-    purposes: [] as const, // ReadonlyArray<PurposeNode>;
-    qualifiers: [] as const, // ReadonlyArray<QualifierNode>;
-    rank: pk, // Scalars["Int"]["output"];
-    requireAdultReservee: true, // Scalars["Boolean"]["output"];
-    requireReservationHandling: false, // Scalars["Boolean"]["output"];
-    reservableTimeSpans: [] as const, // Maybe<ReadonlyArray<ReservableTimeSpanType>>;
-    reservationBlockWholeDay: false, // Scalars["Boolean"]["output"];
-    reservationCancelledInstructionsEn:
-      null /* Maybe< Scalars["String"]["output"] >;*/,
-    reservationCancelledInstructionsFi: null, // Maybe< Scalars["String"]["output"] >;
-    reservationCancelledInstructionsSv: null, // Maybe< Scalars["String"]["output"] >;
-    reservationConfirmedInstructionsEn: null, // Maybe< Scalars["String"]["output"] >;
-    reservationConfirmedInstructionsFi: null, // Maybe< Scalars["String"]["output"] >;
-    reservationConfirmedInstructionsSv: null, // Maybe< Scalars["String"]["output"] >;
-    reservationKind: ReservationKind.DirectAndSeason,
-    reservationPendingInstructionsEn: null, // Maybe<Scalars["String"]["output"]>;
-    reservationPendingInstructionsFi: null, // Maybe<Scalars["String"]["output"]>;
-    reservationPendingInstructionsSv: null, // Maybe<Scalars["String"]["output"]>;
-    reservationStartInterval: ReservationStartInterval.Interval_30Mins,
-    reservationState: ReservationUnitReservationState.Reservable,
-    reservations: null, //Maybe<ReadonlyArray<ReservationNode>>;
-    reservationsMaxDaysBefore: null, // Maybe<Scalars["Int"]["output"]>;
-    reservationsMinDaysBefore: null, // Maybe<Scalars["Int"]["output"]>;
-    resources: [] as const, // ReadonlyArray<ResourceNode>;
-    searchTerms: [] as const, // ReadonlyArray<Scalars["String"]["output"]>;
-    serviceSpecificTerms: null, // Maybe<TermsOfUseNode>;
-    spaces: [] as const, // ReadonlyArray<SpaceNode>;
-    surfaceArea: null, // Maybe<Scalars["Int"]["output"]>;
-    termsOfUseEn: null, // Maybe<Scalars["String"]["output"]>;
-    termsOfUseFi: null, // Maybe<Scalars["String"]["output"]>;
-    termsOfUseSv: null, // Maybe<Scalars["String"]["output"]>;
-    uuid: "dummy-uuid", // Scalars["UUID"]["output"];
-  };
-}
-
-function createMockUnit({ pk }: { pk: number }): UnitNode {
-  return {
-    id: base64encode(`UnitNode:${pk}`),
-    pk, // Maybe<Scalars["Int"]["output"]>;
-    ...generateNameFragment(`Unit ${pk}`),
-    ...generateDescriptionFragment(`Unit Description ${pk}`),
-    email: "", // Scalars["String"]["output"];
-    location: null, // Maybe<LocationNode>;
-    paymentMerchant: null, // Maybe<PaymentMerchantNode>;
-    phone: "", // Scalars["String"]["output"];
-    reservationUnits: [] as const, // ReadonlyArray<ReservationUnitNode>;
-    shortDescriptionEn: `Short description ${pk} EN`, // Scalars["String"]["output"];
-    shortDescriptionFi: `Short description ${pk} FI`, // Scalars["String"]["output"];
-    shortDescriptionSv: `Short description ${pk} SV`, // Scalars["String"]["output"];
-    spaces: [] as const, //; ReadonlyArray<SpaceNode>;
-    tprekId: null, // Maybe<Scalars["String"]["output"]>;
-    unitGroups: [] as const, // ReadonlyArray<UnitGroupNode>;
-    webPage: "", // Scalars["String"]["output"];
-  };
 }
 
 function createMockReservationUnits({
